requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
gspread
google-auth
google-auth-oauthlib
python-dotenv
